do not log aborted error and humanize swap rate error
